requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
requests
openai==2.28.0
pymongo==4.16.0
python-dotenv==1.2.2
numpy==2.4.3
